(2.CustomModelInQML): updated and documented changes on 3-ObjectListModel
