+++ 2.CustomModelsInQML/3-ObjectListModel/main.cpp
@@ -2,7 +2,6 @@
 #include <QQmlContext>
 #include <QQmlApplicationEngine>
 #include <QQuickStyle>
-#include "person.h"
 
 
 int main(int argc, char *argv[])
@@ -11,26 +10,30 @@ int main(int argc, char *argv[])
     QQmlApplicationEngine engine;
     QQuickStyle::setStyle("Material");
 
-    QList<QObject *> personList;
-    personList.append(new Person("John Doe C++","green",32,&engine));
-    personList.append(new Person("Mary Green","blue",23,&engine));
-    personList.append(new Person("Mitch Nathson","dodgerblue",30,&engine));
-    personList.append(new Person("Daniel Sten","red",67,&engine));
-    personList.append(new Person("John Doe C++","green",32,&engine));
-    personList.append(new Person("Mary Green","blue",23,&engine));
-    personList.append(new Person("Mitch Nathson","dodgerblue",30,&engine));
-    personList.append(new Person("Daniel Sten","red",67,&engine));
-    personList.append(new Person("John Doe C++","green",32,&engine));
-    personList.append(new Person("Mary Green","blue",23,&engine));
-    personList.append(new Person("Mitch Nathson","dodgerblue",30,&engine));
-    personList.append(new Person("Daniel Sten","red",67,&engine));
-    personList.append(new Person("John Doe C++","green",32,&engine));
-    personList.append(new Person("Mary Green","blue",23,&engine));
-    personList.append(new Person("Mitch Nathson","dodgerblue",30,&engine));
-    personList.append(new Person("Daniel Sten","red",67,&engine));
+    /**
+     *Setup to expose a moddel as a context property
+     */
 
-    //Expose the list
-    engine.rootContext()->setContextProperty("personModel",QVariant::fromValue(personList));
+    // QList<QObject *> personList;
+    // personList.append(new Person("John Doe C++","green",32,&engine));
+    // personList.append(new Person("Mary Green","blue",23,&engine));
+    // personList.append(new Person("Mitch Nathson","dodgerblue",30,&engine));
+    // personList.append(new Person("Daniel Sten","red",67,&engine));
+    // personList.append(new Person("John Doe C++","green",32,&engine));
+    // personList.append(new Person("Mary Green","blue",23,&engine));
+    // personList.append(new Person("Mitch Nathson","dodgerblue",30,&engine));
+    // personList.append(new Person("Daniel Sten","red",67,&engine));
+    // personList.append(new Person("John Doe C++","green",32,&engine));
+    // personList.append(new Person("Mary Green","blue",23,&engine));
+    // personList.append(new Person("Mitch Nathson","dodgerblue",30,&engine));
+    // personList.append(new Person("Daniel Sten","red",67,&engine));
+    // personList.append(new Person("John Doe C++","green",32,&engine));
+    // personList.append(new Person("Mary Green","blue",23,&engine));
+    // personList.append(new Person("Mitch Nathson","dodgerblue",30,&engine));
+    // personList.append(new Person("Daniel Sten","red",67,&engine));
+
+    // //Expose the list
+    // engine.rootContext()->setContextProperty("personModel",QVariant::fromValue(personList));
 
     const QUrl url(u"qrc:/3-ObjectListModel/main.qml"_qs);
     QObject::connect(&engine, &QQmlApplicationEngine::objectCreated,
