+++ 2.CustomModelsInQML/3-ObjectListModel/main.qml
@@ -50,66 +50,76 @@
 import QtQuick
 import QtQuick.Controls
 import QtQuick.Layouts
+import myModule
 
 Window {
-    id : root
+    id: root
     width: 640
     height: 480
     visible: true
     title: qsTr("Object List Model Demo")
+    PersonList {
+        id: personListModelId
+    }
+    ListView {
+        id: mListView
+        anchors.fill: parent
+        // model : personModel // c++ model exposed as context property in main
+        // model : mModelId // model created in qml
+        // model: DummyListModel{}  // model creaated in qml but exposed from module myModule
+        model: personListModelId.persons // c++ model exposed via module
+        delegate: Component { // refactored delegated exposed via a module and wrapped a QQmlComponent
+            PersonDelegate {
+            }
+        }
+        // setup for an in-place delegate
+        // delegate: Rectangle {
+        //     height: 90
+        //     radius: 10
+        //     color : "gray"
+        //     border.color: "cyan"
+        //     width: root.width
+        //     RowLayout{
+        //         anchors.fill: parent
+        //         anchors.margins: 20
+        //         TextField{
+        //             text : names
+        //             Layout.fillWidth: true
+        //             background: Rectangle{
+        //                 color:"white"
+        //             }
+        //         }
+        //         SpinBox{
+        //             editable: true
+        //             value : age
+        //             Layout.fillWidth: true
+        //             background: Rectangle{
+        //                 color:"white"
+        //             }
+        //         }
+        //         Rectangle{
+        //             width : 50
+        //             height: 50
+        //             color: favoriteColor
 
-       ListView{
-           id : mListView
-           anchors.fill: parent
-           //model : personModel
-           model : mModelId
-           delegate: Rectangle {
-               height: 90
-               radius: 10
-               color : "gray"
-               border.color: "cyan"
-               width: root.width
-               RowLayout{
-                   anchors.fill: parent
-                   anchors.margins: 20
-                   TextField{
-                       text : names
-                       Layout.fillWidth: true
-                       background: Rectangle{
-                           color:"white"
-                       }
-                   }
-                   SpinBox{
-                       editable: true
-                       value : age
-                       Layout.fillWidth: true
-                       background: Rectangle{
-                           color:"white"
-                       }
-                   }
-                   Rectangle{
-                       width : 50
-                       height: 50
-                       color: favoriteColor
+        //         }
+        //     }
+        // }
+    }
 
-                   }
-               }
-           }
-       }
-
-       ListModel{
-           id : mModelId
-           ListElement {
-               names : "Daniel Sten"; favoriteColor : "blue"; age : 10
-           }
-           ListElement {
-               names : "Stevie Wongers"; favoriteColor : "cyan"; age : 23
-           }
-           ListElement {
-               names : "Nicholai Ven"; favoriteColor : "red"; age : 33
-           }
-           ListElement {
-               names : "William Glen"; favoriteColor : "yellowgreen"; age : 45
-           }
-       }
+    // ListModel{
+    //     id : mModelId
+    //     ListElement {
+    //         names : "Daniel Sten"; favoriteColor : "blue"; age : 10
+    //     }
+    //     ListElement {
+    //         names : "Stevie Wongers"; favoriteColor : "cyan"; age : 23
+    //     }
+    //     ListElement {
+    //         names : "Nicholai Ven"; favoriteColor : "red"; age : 33
+    //     }
+    //     ListElement {
+    //         names : "William Glen"; favoriteColor : "yellowgreen"; age : 45
+    //     }
+    // }
 }
