+++ 2.CustomModelsInQML/3-ObjectListModel/module/DummyListModel.qml
@@ -1,0 +1,24 @@
+import QtQuick
+
+ListModel {
+    ListElement {
+        names: "Daniel Sten"
+        favoriteColor: "blue"
+        age: 10
+    }
+    ListElement {
+        names: "Stevie Wongers"
+        favoriteColor: "cyan"
+        age: 23
+    }
+    ListElement {
+        names: "Nicholai Ven"
+        favoriteColor: "red"
+        age: 33
+    }
+    ListElement {
+        names: "William Glen"
+        favoriteColor: "yellowgreen"
+        age: 45
+    }
+}
